Add Darwin and Windows support.
- darwin works, windoze maybe.
