+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/packetdrop/quicwire
+module github.com/nexodus-io/quicwire
 
 go 1.20
 
@@ -37,6 +37,6 @@ require (
 	golang.org/x/exp v0.0.0-20230425010034-47ecfdc1ba53 // indirect
 	golang.org/x/mod v0.10.0 // indirect
 	golang.org/x/net v0.10.0 // indirect
-	golang.org/x/sys v0.8.0 // indirect
+	golang.org/x/sys v0.8.0
 	golang.org/x/tools v0.9.0 // indirect
 )
